+++ lilya/responses.py
@@ -7,6 +7,7 @@ import json
 import os
 import stat
 import typing
+import warnings
 from collections.abc import (
     AsyncIterable,
     Awaitable,
@@ -39,6 +40,7 @@ from lilya.concurrency import iterate_in_threadpool
 from lilya.datastructures import URL, Header
 from lilya.encoders import ENCODER_TYPES, EncoderProtocol, MoldingProtocol, json_encode
 from lilya.enums import Event, HTTPMethod, MediaType
+from lilya.ranges import ContentRanges, parse_range_value
 from lilya.types import Receive, Scope, Send
 
 Content = str | bytes
@@ -287,9 +289,16 @@ class Response:
         prefix = "websocket." if scope["type"] == "websocket" else ""
         await self.resolve_async_content()
         await send(self.message(prefix=prefix))
+        # should be mutation free for both methods
+        mutation_free = "method" in scope and scope["method"].upper() in {
+            HTTPMethod.HEAD,
+            HTTPMethod.OPTIONS,
+        }
+
+        # don't interfere, in case of bodyless requests like head the message is ignored.
         await send({"type": f"{prefix}http.response.body", "body": self.body})
 
-        if self.background is not None:
+        if self.background is not None and not mutation_free:
             await self.background()
 
     def __repr__(self) -> str:
@@ -422,8 +431,6 @@ class StreamingResponse(Response):
                 break
 
     async def stream(self, send: Send) -> None:
-        await send(self.message(prefix=""))
-
         async for chunk in self.body_iterator:
             if not isinstance(chunk, bytes):
                 chunk = chunk.encode(self.charset)
@@ -432,6 +439,17 @@ class StreamingResponse(Response):
         await send({"type": "http.response.body", "body": b"", "more_body": False})
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
+        prefix = "websocket." if scope["type"] == "websocket" else ""
+        await send(self.message(prefix=prefix))
+        # for options and head, we certainly don't want to execute the stream when requesting options
+        send_header_only = "method" in scope and scope["method"].upper() in {
+            HTTPMethod.HEAD,
+            HTTPMethod.OPTIONS,
+        }
+
+        if send_header_only:
+            # no background execution
+            return
         async with anyio.create_task_group() as task_group:
 
             async def wrap(func: Callable[[], Awaitable[None]]) -> None:
@@ -460,13 +478,19 @@ class FileResponse(Response):
         method: str | None = None,
         content_disposition_type: str = "attachment",
         encoders: Sequence[Encoder | type[Encoder]] | None = None,
+        allow_range_requests: bool = True,
     ) -> None:
+        if method:
+            warnings.warn(
+                '"method" parameter is obsolete. It is now automatically deduced.', stacklevel=2
+            )
         self.path = path
         self.status_code = status_code
+        self.allow_range_requests = allow_range_requests
         self.filename = filename
-        self.send_header_only = method is not None and method.upper() == HTTPMethod.HEAD
         if media_type is None:
-            media_type = guess_type(filename or path)[0] or MediaType.TEXT
+            # by default it must be octet
+            media_type = guess_type(filename or path)[0] or MediaType.OCTET
         self.media_type = media_type
         self.background = background
 
@@ -474,6 +498,9 @@ class FileResponse(Response):
             encoder() if isclass(encoder) else encoder for encoder in encoders or _empty
         ]
         self.make_headers(headers)
+
+        if self.allow_range_requests:
+            self.headers["accept-ranges"] = "bytes"
 
         if self.filename is not None:
             content_disposition_filename = quote(self.filename)
@@ -498,6 +525,30 @@ class FileResponse(Response):
         self.headers.setdefault("last-modified", last_modified)
         self.headers.setdefault("etag", etag)
 
+    def check_if_range(self, scope: Scope) -> bool:
+        """Is the if-range matching and the byte ranges are valid?"""
+        received_headers = Header.ensure_header_instance(scope)
+        if_range: str = received_headers.get("if-range", "")
+        # succeeds if_range is matching or empty
+        return not if_range or if_range == cast(str, self.headers["etag"])
+
+    def set_range_headers(self, scope: Scope) -> ContentRanges | None:
+        received_headers = Header.ensure_header_instance(scope)
+        content_ranges = parse_range_value(
+            received_headers.get("range", ""), max_values=int(self.headers["content-length"]) - 1
+        )
+        if content_ranges is None or len(content_ranges.ranges) != 1:
+            # TODO: support multipart ranges
+            return None
+        range_definition = content_ranges.ranges[0]
+
+        self.headers["content-range"] = (
+            f"bytes {range_definition[0]}-{range_definition[1]}/{content_ranges.max_value + 1}"
+        )
+        self.headers["content-length"] = f"{range_definition[1] - range_definition[0] + 1}"
+        self.status_code = status.HTTP_206_PARTIAL_CONTENT
+        return content_ranges
+
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
         if self.stat_result is None:
             try:
@@ -510,35 +561,60 @@ class FileResponse(Response):
                 if not stat.S_ISREG(mode):
                     raise RuntimeError(f"File at path {self.path} is not a file.") from None
 
+        # for options and head, we certainly don't want to send the file when requesting options
+        send_header_only = "method" in scope and scope["method"].upper() in {
+            HTTPMethod.HEAD,
+            HTTPMethod.OPTIONS,
+        }
         prefix = "websocket." if scope["type"] == "websocket" else ""
-        await send(self.message(prefix=prefix))
-        extensions = scope.get("extensions", {})
+        offset: int | None = None
+        if not send_header_only and self.allow_range_requests and self.check_if_range(scope):
+            crange = self.set_range_headers(scope)
+            if crange is not None:
+                # TODO: support multipart ranges
+                offset = crange.ranges[0].start
 
-        if self.send_header_only:
-            await send({"type": "http.response.body", "body": b"", "more_body": False})
-        elif "http.response.pathsend" in extensions:
+        await send(self.message(prefix=prefix))
+
+        if send_header_only:
+            # no background execution
+            return
+
+        extensions = scope.get("extensions", {})
+        if offset is None and "http.response.pathsend" in extensions:
             await send({"type": "http.response.pathsend", "path": self.path})
         elif "http.response.zerocopysend" in extensions:
             async with await anyio.open_file(self.path, mode="rb") as file:
                 size = int(self.headers["content-length"])
+                if offset:
+                    await file.seek(offset)
                 more_body = True
                 while more_body:
                     more_body = size > self.chunk_size
+                    count = self.chunk_size
+                    if not more_body:
+                        count = size
                     await send(
                         {
                             "type": "http.response.zerocopysend",
                             "file": file.fileno(),  # type: ignore
-                            "count": self.chunk_size,
+                            "count": count,
                             "more_body": more_body,
                         }
                     )
-                    size -= self.chunk_size
+                    size -= count
         else:
             async with await anyio.open_file(self.path, mode="rb") as file:
+                size = int(self.headers["content-length"])
+                if offset:
+                    await file.seek(offset)
                 more_body = True
                 while more_body:
-                    chunk = await file.read(self.chunk_size)
-                    more_body = len(chunk) == self.chunk_size
+                    more_body = size > self.chunk_size
+                    count = self.chunk_size
+                    if not more_body:
+                        count = size
+                    chunk = await file.read(count)
                     await send(
                         {
                             "type": "http.response.body",
@@ -546,6 +622,7 @@ class FileResponse(Response):
                             "more_body": more_body,
                         }
                     )
+                    size -= count
         if self.background is not None:
             await self.background()