+++ tests/test_responses.py
@@ -11,6 +11,8 @@ import pytest
 from lilya import status
 from lilya.apps import Lilya
 from lilya.background import Task
+from lilya.compat import md5_hexdigest
+from lilya.datastructures import Header
 from lilya.encoders import Encoder
 from lilya.middleware import DefineMiddleware
 from lilya.middleware.sessions import SessionMiddleware
@@ -297,6 +299,10 @@ def test_file_response(tmpdir, test_client_factory):
 
     assert filled_by_bg_task == ""
     client = test_client_factory(app)
+    response = client.head("/")
+    expected_disposition = 'attachment; filename="example.png"'
+    assert response.status_code == status.HTTP_200_OK
+    assert response.content == b""
     response = client.get("/")
     expected_disposition = 'attachment; filename="example.png"'
     assert response.status_code == status.HTTP_200_OK
@@ -324,8 +330,8 @@ async def test_file_response_optimizations(tmpdir, extensions, result, anyio_bac
 
     fresponse = FileResponse(path=path, filename="example.png")
     fresponse.chunk_size = 10
-    responses = Queue()
-    await fresponse({"extensions": extensions, "type": "response"}, None, responses.put)
+    responses: Queue[typing.Any] = Queue()
+    await fresponse({"extensions": extensions, "type": "http.request"}, None, responses.put)
     response1 = await responses.get()
     response2 = await responses.get()
 
@@ -342,6 +348,120 @@ async def test_file_response_optimizations(tmpdir, extensions, result, anyio_bac
         while response2["more_body"]:
             response2 = await responses.get()
         assert not response2["more_body"]
+
+
+@pytest.mark.parametrize(
+    "extensions,extpath",
+    [
+        ({}, "none"),
+        ({"http.response.pathsend": {}}, "http.response.pathsend"),
+        ({"http.response.zerocopysend": {}, "http.response.pathsend": {}}, "both"),
+        ({"http.response.zerocopysend": {}}, "http.response.zerocopysend"),
+    ],
+)
+@pytest.mark.parametrize("matching_ifrange", [True, False, None])
+@pytest.mark.parametrize(
+    "byterange,start,end",
+    [
+        ("bytes=1-10", 1, 10),
+        ("bytes=1-", 1, 999),
+        ("bytes=0", 0, 0),
+        ("bytes=-10", 989, 999),
+    ],
+)
+async def test_file_response_byte_range(
+    tmpdir, extensions, extpath, byterange, start, end, matching_ifrange, anyio_backend
+):
+    path = os.path.join(tmpdir, "xyz")
+    content = os.urandom(1000)
+    with open(path, "wb") as file:
+        file.write(content)
+    stat_result = os.stat(path)
+    if matching_ifrange:
+        etag_base = str(stat_result.st_mtime) + "-" + str(stat_result.st_size)
+        ifrange = md5_hexdigest(etag_base.encode(), usedforsecurity=False)
+    elif matching_ifrange is None:
+        ifrange = ""
+    else:
+        ifrange = "123456"
+
+    fresponse = FileResponse(path=path, filename="example.png")
+    responses: Queue[typing.Any] = Queue()
+
+    async def put(message: typing.Any) -> None:
+        if "file" in message:
+            rob = await anyio.open_file(message["file"], "rb", closefd=False)
+            message["body"] = await rob.read(message["count"])
+        await responses.put(message)
+
+    await fresponse(
+        {
+            "extensions": extensions,
+            "type": "http.request",
+            "headers": [("range", byterange), ("if-range", ifrange.encode())],
+        },
+        None,
+        put,
+    )
+    response1 = await responses.get()
+    headers = Header.from_scope(response1)
+    response2 = await responses.get()
+    assert headers["accept-ranges"] == "bytes"
+    # test only one response, don't want to merge for tests
+    if "path" not in response2:
+        assert not response2["more_body"]
+    if matching_ifrange or matching_ifrange is None:
+        assert int(headers["content-length"]) == end - start + 1
+        if extpath == "http.response.zerocopysend" or extpath == "both":
+            assert response2["body"] == content[start : end + 1]
+        else:
+            assert response2["body"] == content[start : end + 1]
+    else:
+        assert int(headers["content-length"]) == len(content)
+        if extpath == "http.response.pathsend" or extpath == "both":
+            assert response2["path"] == path
+        elif extpath == "http.response.zerocopysend":
+            assert response2["count"] == int(headers["content-length"])
+            assert response2["body"] == content
+        else:
+            assert response2["body"] == content
+
+
+@pytest.mark.parametrize(
+    "byterange",
+    [
+        ("megabytes=1-10"),
+        ("bytes=1-10, 1-1"),
+        ("bytes=100-10"),
+        # not supported yet, despite no error
+        ("bytes=1-10, 10-29"),
+    ],
+)
+async def test_file_response_byte_range_error(tmpdir, byterange, anyio_backend):
+    path = os.path.join(tmpdir, "xyz")
+    content = os.urandom(1000)
+    with open(path, "wb") as file:
+        file.write(content)
+
+    fresponse = FileResponse(path=path, filename="example.png")
+    responses: Queue[typing.Any] = Queue()
+    await fresponse(
+        {
+            "extensions": {},
+            "type": "http.request",
+            "headers": [("range", byterange)],
+        },
+        None,
+        responses.put,
+    )
+    response1 = await responses.get()
+    headers = Header.from_scope(response1)
+    response2 = await responses.get()
+    assert headers["accept-ranges"] == "bytes"
+    assert int(headers["content-length"]) == len(content)
+    assert "range" not in headers
+    assert not response2["more_body"]
+    assert response2["body"] == content
 
 
 def test_file_response_with_directory_raises_error(tmpdir, test_client_factory):
@@ -635,10 +755,38 @@ def test_populate_headers(test_client_factory):
 
 
 def test_head_method(test_client_factory):
-    app = Response("hello, world", media_type="text/plain")
+    executed: bool = False
+
+    async def hello() -> str:
+        nonlocal executed
+        executed = True
+        return "hello, world"
+
+    app = Response(hello(), media_type="text/plain")
+    assert not executed
     client = test_client_factory(app)
     response = client.head("/")
     assert response.text == ""
+    # is executed anyway for content-length
+    assert executed
+
+
+def test_options_method(test_client_factory):
+    executed: bool = False
+
+    async def hello() -> str:
+        nonlocal executed
+        executed = True
+        return "hello, world"
+
+    app = Response(hello(), media_type="text/plain")
+    assert not executed
+    client = test_client_factory(app)
+    response = client.options("/")
+    # body is allowed for options
+    assert response.text == "hello, world"
+    # is executed anyway for content-length
+    assert executed
 
 
 def test_empty_response(test_client_factory):
@@ -717,5 +865,5 @@ async def test_streaming_response_stops_if_receiving_http_disconnect():
     response = StreamingResponse(content=stream_indefinitely())
 
     with anyio.move_on_after(1) as cancel_scope:
-        await response({}, receive_disconnect, send)
+        await response({"type": "http"}, receive_disconnect, send)
     assert not cancel_scope.cancel_called, "Content streaming should stop itself."